v1.30.0
-- Simplify code which assigns labels to data plot axes in MainGUI by using object oriented class methods to access objects' data and fix an issue in the ppf check which decides whether to add length units to y-axis label.

a) Expand implementation of GetVar() method to other DataTree classes like FuncCallClass in order to be able to search for user-settable parameters in a way that keeps the implementation details from the client application which is MainGUI (or any other GUI).

b) Also implement GetVar() method for SubjClass and GroupClass in order to search for run-level user-settable parameters. Each of these methods calls common code implemented in GetVar() method of TreeNodeClass while code specific to group or subject if implemented in each classes GetVar method.

c) Implement GetLengthUnit() method in Snirf/SnirfClass.m to allow GUIs to get information about parameter units if it exists.

d) Fix ppf check which decides whether to add length units to y-axis label. It fails when selecting multiple wavelengths or multiple Hb types. Also it is not correct to use iWl (currently selected wavelength) for indexing to ppf because you're plotting concentration NOT OD or raw data so the currently selected wavelength is irrelevant. All you need to know is did ppf == 1 for ANY wavelength when calculating concentration. If yes then add length unit to the concentration unit label.