Fix on group averaging
When a subject had nan values on a channel/sbj, the group result ended being nan, even though the rest of the sbjs are good. Homer2 was unconsciously dealing with this issue by pruning outliers from group, since we do not have that feature in homer3, the issue came up. Issue is fixed now by identifying those channels and removing from group averaging.