Update hmrG_SubjAvg.m
FIX; When a subject had nan values on a channel/sbj, the group result ended being nan, even though the rest of the sbjs are good. Homer2 was unconsciously dealing with this issue by pruning outliers from group with a preset threshold. Since we do not have that feature in homer3, the issue came up. Issue is fixed now by identifying those channels with nan and removing them from group averaging.