Update DisplayDataConc.m
Condition index was missing from dHbMlStd calculation, now fixed.
Also now plotting standard deviation not error. Reason for change: only run level calculation was there for standard error, and session and group avg was not getting calculated properly. In the future, if we want to have session SME, we would need to input nTrialRuns and sum trials for all runs, and same goes for group SME.