Critical bug fix SplineSG mlAct
SplineSG has mlActAuto as an input. Due to the changes in homer3, mlActAuto containes only the first wavelength. SplineSG thus failed to clean the data from the second wavelength.
Note for homer2 users: This was working properly for homer2.
Note for homer3 users: Please reconsider running your pipeline with this version and up.